+++ src/__tests__/array.test.js
@@ -1,6 +1,6 @@
 // @flow strict
 
-import { array } from '../array';
+import { array, nonEmptyArray } from '../array';
 import { guard } from '../guard';
 import { number } from '../number';
 import { object } from '../object';
@@ -39,5 +39,23 @@ describe('array', () => {
 
         const g2 = guard(array(object({ name: string })));
         expect(() => g2([{ name: 123 }])).toThrow('^ index 0');
+    });
+});
+
+describe('nonEmptyArray', () => {
+    const strings = guard(nonEmptyArray(string));
+    const numbers = guard(nonEmptyArray(number));
+
+    it('works like normal array', () => {
+        expect(strings(['foo', 'bar'])).toEqual(['foo', 'bar']);
+        expect(numbers([1, 2, 3])).toEqual([1, 2, 3]);
+
+        expect(() => strings([1])).toThrow('Must be string');
+        expect(() => numbers(['foo'])).toThrow('Must be number');
+    });
+
+    it('but empty array throw, too', () => {
+        expect(() => strings([])).toThrow('Must be non-empty array');
+        expect(() => numbers([])).toThrow('Must be non-empty array');
     });
 });